:recycle: [EditItem] Rename add item screen
