+++ src/components/add-edit-item/container/index.js
@@ -2,7 +2,7 @@ import * as ReactRedux from 'react-redux'
 import itemAction from 'redux/actions/item'
 import selectors from 'redux/selectors'
 
-import Add from '..'
+import AddEditItem from '..'
 
 const mapStateToProp = state => ({
   items: selectors.items.allOrdered(state),
@@ -15,4 +15,4 @@ const mapDispatchToProps = dispatch => ({
 export default ReactRedux.connect(
   mapStateToProp,
   mapDispatchToProps
-)(Add)
+)(AddEditItem)
