+++ src/components/routes/index.js
@@ -10,7 +10,7 @@ import * as Reselect from 'reselect'
 import NavBar from 'components/nav-bar'
 import Store from 'components/store/container'
 import Sales from 'components/sales'
-import AddEditItem from 'components/add/container'
+import AddEditItem from 'components/add-edit-item/container'
 import Payment from 'components/payment'
 import Home from 'components/onboarding/home'
 import GetWallet from 'components/onboarding/get-wallet'
